240820 수업 준비

+++ src/main/java/org/example/repository/book/jpa/JpaBookRepository.java
@@ -33,13 +33,9 @@ public class JpaBookRepository {
         return book;
     }
 
-    public int delete(Long id) {
+    public void delete(Long id) {
         Book book = em.find(Book.class, id);
-        if (book != null) {
-            em.remove(book);
-            return 1;
-        }
-        return 0;
+        if (book != null) em.remove(book);
     }
 
 }
